X-O-X_game
This game is a tile game played by one player against the computer, the goal being to line up three "X" signs in a row or prevent the computer from winning.
This game is a tile game played by one player against the computer, the goal being to line up three "X" signs in a row or prevent the computer from winning.